refactoring footer
